projects/OCR/requirements.txt
opencv-python
PyQt5
requests
google-generativeai
